pyproject.toml
[project]
name = "xiaohongshu-smart-gen"
version = "1.0.0"
description = "小红书多垂类内容智能生成工具 - 跨平台 Python 实现"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "talhuang1231" }
]

dependencies = [
    "requests>=2.28.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0",
    "pytest-cov>=6.0",
]

[project.scripts]
xhs-gen = "scripts.xhs_cli.cli:main"
xhs-do = "scripts.xhs_cli.cli:main_do"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E402", "RUF001", "RUF002", "RUF003"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
